+++ computers/src/main/java/dev/jstech/computers/os/ProgramVersions.java
@@ -41,6 +41,7 @@ public final class ProgramVersions {
             Map.entry("mirror", "2.5"),
             Map.entry("screenfetch", "3.9.1"),
             Map.entry("sgsc", "1.0"),
+            Map.entry("scc", "1.0"),
             Map.entry("sigma", "1.0"),
             Map.entry("virtual_studio", "17.0"),
             Map.entry("virtual_studio_code", "1.85"),